Mini Assignment 01/10
Turning in the "Hello World!" Mini Assignment. I first listed the "Hello, World!" message and four different replies as objects. Then, I started the count repetition and in the last section created the loop with conditional statements. I ran each of the objects first to create the values, then the count, and then the loop. I included informative comments either above or besides the commands. The outcome is a text with Hello, world! being repeated in an intercalated way with the four other messages until the loop stops/breaks. @nrdowling @hjennybo