Wrote in the Data
First, I wrote in my original data spreadsheet with the readr function for excel. Then, as a test run, I wrote in a test csv file to see if the function was working. Externally, created a separate spreadsheet of my data in Excel by saving as a csv file. But I had trouble with the readr csv function because my dataset was too large. I'll also write this down in an issue. @nrdowling @hjennybo