CCRollCallREG Data Upload
This spreadsheet contains all of the Roll Call Votes in the Regulatory Phase of the Chilean Constitutional Convention. The data was translated from Spanish to English, the variables were turned into binary categories where a vote in favor= 1; a vote against= 0; a no vote or abstention= blank. Also all of the articles voted on were "anonymized" for simplicity so they're expressed as reg_1, reg_2, reg_3....reg_1009.